Added inheritance relation
Modify Staff and Doctor classes as the subclass of Person and inherit its method and common variable thru overriding
